==> syntax/03_UnionEstadoDesaparicion.ipynb <==
import pandas as pd
archivo1 = pd.read_csv('C:\\Users\\Braulio\\Documents\\Braulio\\tot_reg_desap.csv', encoding='latin1')
archivo2 = pd.read_csv('C:\\Users\\Braulio\\Documents\\Braulio\\RNPDNO_V5.csv', encoding='latin1')

archivo1.columns

archivo1

archivo2

archivo2_ordenado = archivo2.sort_values(by='IDX_REPORTE')
archivo2_ordenado

archivo1_ordenado = archivo1.sort_values(by='DC_id')
archivo1_ordenado

def extract_key(idx):
    parts = idx.split('-')
    if len(parts) > 1:
        return parts[1].lstrip('0')
    return idx

archivo2['Clave'] = archivo2['IDX_REPORTE'].apply(extract_key)
archivo2

archivo1['Clave'] = archivo1['DC_id'].apply(lambda x: str(x).split()[0] if pd.notnull(x) else '')
archivo1

print(archivo2[['IDX_REPORTE', 'Clave']])

print(archivo1[['DC_id', 'Clave']])

archivo1_s = archivo1[['Clave', 'fecha_desap', 'estatus_busqueda']]
data = archivo2.merge(archivo1_s, on='Clave', how='left')
data
